REF/WIP: first pass re-organizing files, redirecting ui file paths
