+++ pmpsui/templates/arbiter_outputs_entry.ui
