+++ pmpsui/templates/arbiter_outputs_header.ui
