+++ pmpsui/ui/arbiter_outputs.ui
@@ -171,8 +171,20 @@
         <property name="frameShape">
          <enum>QFrame::NoFrame</enum>
         </property>
+        <property name="macros" stdset="0">
+         <string>{}</string>
+        </property>
         <property name="filename" stdset="0">
-         <string>templates/arbiter_outputs_header.ui</string>
+         <string>../templates/arbiter_outputs_header.ui</string>
+        </property>
+        <property name="loadWhenShown" stdset="0">
+         <bool>true</bool>
+        </property>
+        <property name="disconnectWhenHidden" stdset="0">
+         <bool>true</bool>
+        </property>
+        <property name="followSymlinks" stdset="0">
+         <bool>false</bool>
         </property>
        </widget>
       </item>
@@ -186,7 +198,7 @@
           <rect>
            <x>0</x>
            <y>0</y>
-           <width>648</width>
+           <width>672</width>
            <height>185</height>
           </rect>
          </property>
