+++ pmpsui/ui/line_beam_parameters.ui
@@ -2867,12 +2867,12 @@ and Zero BC</string>
   <customwidget>
    <class>EvByteIndicator</class>
    <extends>PyDMByteIndicator</extends>
-   <header>widgets</header>
+   <header>pmpsui.widgets</header>
   </customwidget>
   <customwidget>
    <class>BCByteIndicator</class>
    <extends>PyDMByteIndicator</extends>
-   <header>widgets</header>
+   <header>pmpsui.widgets</header>
   </customwidget>
  </customwidgets>
  <resources/>
